Changes to Moisture Sensor

+++ main/AgriSensor/lib/Moisture_Sensor/Moisture_Sensor.h
@@ -9,7 +9,7 @@ class MoistureSensor
 
     MoistureSensor(int pin);
     void setup();
-    float read();
+    int read();
     void calibrate();
   private:
     int _pin;
